Move xml utility functions out of parser.cpp and into their own xmlutils compilation unit.
